fix to last commit

+++ client/src/world.c
@@ -111,7 +111,7 @@ void World_QueueChunk(Chunk *chunk) {
 
 void World_AddChunk(Vector3 position) {
 
-    long int p = (long)((int)(position.x)&4095)<<20 | ((int)(position.z)&4095)<<8 | ((int)(position.y)&255);
+    long int p = (long)((int)(position.x)&4095)<<20 | (long)((int)(position.z)&4095)<<8 | ((int)(position.y)&255);
     int index = hmgeti(world.chunks, p);
     if(index == -1) {
         //Add chunk to list
@@ -128,7 +128,7 @@ void World_AddChunk(Vector3 position) {
 }
 
 Chunk* World_GetChunkAt(Vector3 position) {
-    long int p = (long)((int)(position.x)&4095)<<20 | ((int)(position.z)&4095)<<8 | ((int)(position.y)&255);
+    long int p = (long)((int)(position.x)&4095)<<20 | (long)((int)(position.z)&4095)<<8 | ((int)(position.y)&255);
     int index = hmgeti(world.chunks, p);
     if(index >= 0) {
         return world.chunks[index].value;
@@ -138,7 +138,7 @@ Chunk* World_GetChunkAt(Vector3 position) {
 }
 
 void World_RemoveChunk(Chunk *curChunk) {
-    long int p = (long)((int)(curChunk->position.x)&4095)<<20 | ((int)(curChunk->position.z)&4095)<<8 | ((int)(curChunk->position.y)&255);
+    long int p = (long)((int)(curChunk->position.x)&4095)<<20 | (long)((int)(curChunk->position.z)&4095)<<8 | ((int)(curChunk->position.y)&255);
 
     int index = hmgeti(world.chunks, p);
     if(index >= 0) {
